New version: gsw_geo_strf_dyn_height_1.c
This is a complete re-implementation of the basic geostrophic
streamfunction calculation.

The debug_makefile now makes the test_strf_1 test program for this
version.  Both are suitable for running with valgrind to check
for memory errors.

This function differs from gsw_geo_strf_dyn_height in several ways:

1) Most importantly, to limited the extent that I have tested it,
it does not leak memory, use uninitialized memory, or otherwise
behave badly, unlike the original.

2) It is using simple linear interpolation.  A fancier routine
could be plugged in if desired.  Alternatively, it might make
more sense to leave fancier interpolation to the calling code.
Note that the original (and ancient) RR68 algorithm, for which
I think the implementation is buggy, is no longer used in
GSW-Matlab.  I don't think it is worth the effort it would take
to debug it.

3) It requires the reference pressure to be within the range of the
profile; the original allowed it to be less than the minimum pressure,
which has no use case that I can see, but added complexity to the code.

4) It returns an integer error code instead of either the input
dynamic height pointer or NULL.  This makes it easy for the calling
routine to safely free memory in the event of an error.